backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-pptx==0.6.23
openai==1.10.0
anthropic==0.18.0
google-generativeai==0.3.2
httpx==0.26.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
# Document conversion libraries (commented out due to installation issues on some systems)
# pdf2docx==0.5.8
# pdf2pptx==1.0.5
# PyMuPDF==1.24.3
aspose-slides==24.1.0
aspose-words==24.1.0